backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
fastapi==0.123.9
google-auth==2.43.0
google-genai==1.53.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
numpy==2.3.5
portalocker==3.2.0
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311; sys_platform == "win32"
qdrant-client==1.16.1
requests==2.32.5
rsa==4.9.1
starlette==0.50.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
